+++ app/src/main/kotlin/io/github/bsfranca2/athena/dto/issue/BugIssueDto.kt
@@ -4,15 +4,15 @@ import io.github.bsfranca2.athena.enum.IssueType
 import java.time.LocalDateTime
 
 class BugIssueDto(
-        id: Int,
+        id: Long,
         title: String,
         description: String,
         status: String,
         priority: Int,
         startDate: LocalDateTime?,
         endDate: LocalDateTime?,
-        val parentId: Int? = null,
+        val parentId: Long? = null,
         val children: MutableList<IssueDto> = mutableListOf(),
-        val assignedTo: MutableList<Int> = mutableListOf(),
-        createdBy: Int
+        val assignedTo: MutableList<Long> = mutableListOf(),
+        createdBy: Long
 ) : IssueDto(id, IssueType.BUG, title, description, status, priority, startDate, endDate, createdBy)
