+++ app/src/main/kotlin/io/github/bsfranca2/athena/dto/issue/TimeEntryDto.kt
@@ -5,14 +5,14 @@ import javax.validation.constraints.Min
 import javax.validation.constraints.NotNull
 
 data class TimeEntryDto(
-        val id: Int = -1,
+        val id: Long = -1,
         @field:Min(1)
-        val issue: Int = -1,
+        val issue: Long = -1,
         val description: String = "",
         val registerAt: LocalDateTime = LocalDateTime.now(),
         val timeSpent: Int = 0,
         @field:Min(1)
-        val createdBy: Int = -1,
+        val createdBy: Long = -1,
         val createdAt: LocalDateTime = LocalDateTime.now(),
         val updatedAt: LocalDateTime? = null
 )
