+++ app/src/main/kotlin/io/github/bsfranca2/athena/exception/IssueNotFoundException.kt
@@ -1,3 +1,3 @@
 package io.github.bsfranca2.athena.exception
 
-data class IssueNotFoundException(val id: Int) : Throwable("Issue $id not found")
+data class IssueNotFoundException(val id: Long) : Throwable("Issue $id not found")
