+++ app/src/main/kotlin/io/github/bsfranca2/athena/repository/IssueRepository.kt
@@ -4,7 +4,7 @@ import io.github.bsfranca2.athena.entity.Issue
 import io.github.bsfranca2.athena.entity.User
 import org.springframework.data.repository.CrudRepository
 
-interface IssueRepository : CrudRepository<Issue, Int> {
+interface IssueRepository : CrudRepository<Issue, Long> {
     fun findByCreatedBy(createdBy: User): MutableList<Issue>
 
     fun findByCreatedByAndParentIsNull(createdBy: User): MutableList<Issue>
