+++ app/src/main/kotlin/io/github/bsfranca2/athena/repository/PrivilegeRepository.kt
@@ -3,6 +3,6 @@ package io.github.bsfranca2.athena.repository
 import io.github.bsfranca2.athena.entity.Privilege
 import org.springframework.data.repository.CrudRepository
 
-interface PrivilegeRepository : CrudRepository<Privilege, Int> {
+interface PrivilegeRepository : CrudRepository<Privilege, Long> {
     fun findByName(name: String?): Privilege?
 }
