+++ app/src/main/kotlin/io/github/bsfranca2/athena/repository/RoleRepository.kt
@@ -3,6 +3,6 @@ package io.github.bsfranca2.athena.repository
 import io.github.bsfranca2.athena.entity.Role
 import org.springframework.data.repository.CrudRepository
 
-interface RoleRepository : CrudRepository<Role, Int> {
+interface RoleRepository : CrudRepository<Role, Long> {
     fun findByName(name: String?): Role?
 }
